Fix: 예외처리 추가 및 컨벤션 수정
