+++ postings/views/__init__.py
@@ -1,4 +1,5 @@
-from .view_company     import *
-from .view_companylist import *
-from .view_job_posting import *
-from .view_menu        import *
+from .view_company_list import *
+from .view_company      import *
+from .view_job_list     import *
+from .view_job_posting  import *
+from .view_menu         import *
