+++ requirements.txt
@@ -19,6 +19,7 @@ django-cors-headers==3.7.0
 django-cors-middleware==1.5.0
 django-environ==0.4.5
 django-extensions==3.1.3
+django-rest-swagger==2.2.0
 django-storages==1.11.1
 django-tinymce4-lite==1.8.0
 djangorestframework==3.12.4
@@ -36,6 +37,7 @@ MarkupSafe==2.0.1
 matplotlib-inline==0.1.2
 mysqlclient==2.0.3
 oauthlib==3.1.1
+openapi-codec==1.3.2
 packaging==20.9
 parso==0.8.2
 pexpect==4.8.0
@@ -54,6 +56,7 @@ requests-oauthlib==1.3.0
 ruamel.yaml==0.17.10
 ruamel.yaml.clib==0.2.4
 s3transfer==0.4.2
+simplejson==3.17.2
 six==1.16.0
 social-auth-core==4.1.0
 sqlparse==0.4.1
