+++ users/models/model_university.py
@@ -8,11 +8,13 @@ class University(models.Model):
 
 
 class UserUniversity(models.Model):
-    start_date  = models.DateTimeField()
-    end_date    = models.DateTimeField()
-    is_enrolled = models.BooleanField()
-    major       = models.CharField(max_length = 45)
-    description = models.CharField(max_length = 1000)
+    start_date   = models.DateTimeField(null = True)
+    end_date     = models.DateTimeField(null = True)
+    is_enrolled  = models.BooleanField(default = False)
+    major        = models.CharField(max_length = 45, null = True)
+    description  = models.CharField(max_length = 1000, null = True)
+    general_user = models.ForeignKey('users.GeneralUser', on_delete = models.CASCADE)
+    university   = models.ForeignKey('users.University', on_delete = models.SET_NULL, null = True)
 
     class Meta:
         db_table = 'user_university'
